moved clear logic to `manyToOneRelation` field
