+++ assets/js/src/core/modules/widget-editor/components/widget-type-form/components/specific-panel/specific-panel.tsx
@@ -12,20 +12,13 @@ import { Form } from '@Pimcore/components/form/form'
 import { FormKit } from '@Pimcore/components/form/form-kit'
 import { Switch } from '@Pimcore/components/switch/switch'
 import { InputNumber } from 'antd'
-import React, { useEffect } from 'react'
+import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { ElementTypeSelect } from '../../../element-type-select/element-type-select'
 import { ManyToOneRelation } from '../many-to-one-relation/many-to-one-relation'
-import { useWidgetFormContext } from '../../context/hooks/use-widget-form-context'
 
 export const SpecificPanel = (): React.JSX.Element => {
   const { t } = useTranslation()
-  const { form } = useWidgetFormContext()
-  const elementType = Form.useWatch('elementType', form)
-
-  useEffect(() => {
-    form.setFieldValue('rootFolder', null)
-  }, [elementType, form])
 
   return (
     <FormKit.Panel
@@ -44,7 +37,7 @@ export const SpecificPanel = (): React.JSX.Element => {
         label={t('widget-editor.widget-form.specific.root-folder')}
         name="rootFolder"
       >
-        <ManyToOneRelation elementType={elementType} />
+        <ManyToOneRelation />
       </Form.Item>
 
       <Form.Item
